Trait ckks_he::algebra::big_int::BigInt

pub trait BigInt: Sized + Add<Output = Self> + Sub<Output = Self> + Neg<Output = Self> + Mul<Output = Self> + Div<Output = Self> + Rem<Output = Self> + for<'a> Add<&'a Self, Output = Self> + for<'a> Sub<&'a Self, Output = Self> + for<'a> Mul<&'a Self, Output = Self> + for<'a> Div<&'a Self, Output = Self> + for<'a> Rem<&'a Self, Output = Self> + Copy + Default + Display + Debug + From<i64> + FromFloat + ToFloat + Eq + PartialOrd + Ord + UniformSamplable + Zero {
    // Provided methods
    fn new(value: i64) -> Self { ... }
    fn modulo(&self, modulus: Self) -> RingMod<Self> { ... }
    fn fast_exp(&self, exponent: u32) -> Self { ... }
    fn remainder(&self, other: &Self) -> Self { ... }
}

A trait representing large integers with common arithmetic operations.

Provided Methods

fn new(value: i64) -> Self

fn modulo(&self, modulus: Self) -> RingMod<Self>

Reduce an integer modulo a modulus

fn fast_exp(&self, exponent: u32) -> Self

Performs fast exponentiation of the value to the given exponent.

Examples

use ckks_he::algebra::big_int::BigInt;
use bnum::types::I1024;

let exponent = 9;
let big_int_value = I1024::from(2 as i64);

let result = big_int_value.fast_exp(exponent);
assert_eq!(result, I1024::from(512 as i64)); // 2^9

fn remainder(&self, other: &Self) -> Self

Picks a representative of self modulo other in the range ]-other/2, other/2].

Assumes other is greater than 0.

Examples

use ckks_he::algebra::big_int::BigInt;
use bnum::types::I1024;

let big_int_value = I1024::from(38);
let big_int_other = I1024::from(10);

let remainder = big_int_value.remainder(&big_int_other);
assert_eq!(remainder, I1024::from(-2)); // 38 % 10 == -2