Struct ckks_he::encoding::Encoder

pub struct Encoder<T: BigInt> {
    pub scaling_factor: f64,
    /* private fields */
}

Fields
scaling_factor: f64
Implementations

impl<T: BigInt> Encoder<T>

pub fn encode_to_integer(&self, plaintext: &[C64]) -> CyclotomicRing<T>

pub fn encode(&self, plaintext: &[C64]) -> CyclotomicRing<RingMod<T>>

pub fn encode_no_rescale(&self, plaintext: &[C64]) -> CyclotomicRing<RingMod<T>>

pub fn decode_integer(&self, message: CyclotomicRing<T>) -> Vec<C64>

pub fn decode_no_rescale( &self, message: &CyclotomicRing<RingMod<T>> ) -> Vec<C64>

pub fn decode(&self, message: &CyclotomicRing<RingMod<T>>) -> Vec<C64>

impl<T: BigInt> Encoder<T>

Encoder with functions encode to go from plaintext to ciphertext ring and decode from ciphertext ring to plaintext. The dimension of the ciphertext ring is 2^h with h = dimension_exponent. The ciphertext ring is a cyclotomic ring $Z/qZ [X] / (1 + X^{2^h})$. Here q = modulus. The space of plaintext is a complex vector space of dimension 2^{h-1}.

pub fn new(dimension_exponent: u32, modulus: T, scaling_factor: f64) -> Self

Generate an encoder for the space of ciphertexts $\ZZ/(modulus \ZZ) [X] / (1 + X ^{2^{dimension_exponent}})$ Decoding $decode: \ZZ/(modulus \ZZ) [X] / (1 + X ^{2^{dimension_exponent}}) \to \CC^{2^{dimension_exponent-1}}$ is the composition $decode = \pi\inv \circ \sigma\inv$ with $\sigma$ the canonical embedding. In the encoder $\sigma$ and $\sigma\inv$ are stored as matrices of roots of unity.