Struct ckks_he::keys::client_key::ClientKey

pub struct ClientKey<T: BigInt> {
    pub encoder: Encoder<T>,
    pub secret_key: SecretKey<T>,
    pub public_key: PublicKey<T>,
    pub plaintext_dimension: usize,
}

Encapsulates the data necessary for the client side, including the secret key, public key, and an encoder.

The ClientKey struct is responsible for managing the encryption and decryption processes in the homomorphic encryption scheme. It holds the secret key for decryption, the public key for encryption, and an encoder for encoding and decoding plaintext messages. This struct provides methods for encrypting plaintext messages, decrypting ciphertexts, and calculating the expected error after decryption.

Examples

Basic usage:

use ckks_he::keys::key_generator::generate_pair_keys_toy;
use ckks_he::random_distributions::generate_random_vector;
use ckks_he::keys::client_key::{calculate_error, to_plaintext};

let (mut client_key, server_key) = generate_pair_keys_toy();
let plaintext_dimension = client_key.plaintext_dimension();
let plaintext_bound = 5.0;
let real_plaintext = generate_random_vector(
    plaintext_dimension,
    -plaintext_bound,
    plaintext_bound,
);
let plaintext = to_plaintext(&real_plaintext);
let ciphertext = client_key.encrypt(&plaintext, plaintext_bound).unwrap();
let decrypted = client_key.decrypt(&ciphertext);
let error = calculate_error(&plaintext, &decrypted);
assert!(error < client_key.rescaled_error(&ciphertext), "Error larger than expected");

Fields
encoder: Encoder<T>secret_key: SecretKey<T>public_key: PublicKey<T>plaintext_dimension: usize
Implementations

impl<T: BigInt> ClientKey<T>

pub fn new( parameters: KeyGenerationParameters<T>, secret_key: SecretKey<T>, public_key: PublicKey<T> ) -> Self

pub fn encrypt( &mut self, clear: &Plaintext, upper_bound: f64 ) -> Result<Ciphertext<T>, EncryptionError>

Encrypt a plaintext, an upperbound must be given.

Examples

use ckks_he::keys::key_generator::generate_pair_keys_toy;
use ckks_he::random_distributions::generate_random_vector;
use ckks_he::keys::client_key::{calculate_error, to_plaintext};

let (mut client_key, server_key) = generate_pair_keys_toy();
let plaintext_dimension = 1 << (client_key.dimension_exponent() -1);
let plaintext_bound = 5.0;
let real_plaintext = generate_random_vector(
        plaintext_dimension,
        -plaintext_bound,
        plaintext_bound,
);
let plaintext = to_plaintext(&real_plaintext);
let ciphertext = client_key.encrypt(&plaintext, plaintext_bound).unwrap();
let decrypted = client_key.decrypt(&ciphertext);
let error = calculate_error(&plaintext, &decrypted);
assert!(error < client_key.rescaled_error(&ciphertext), "Error large than expected");

pub fn decrypt(&mut self, ct: &Ciphertext<T>) -> Plaintext

Decrypt a ciphertext.

Examples

use ckks_he::keys::key_generator::generate_pair_keys_toy;
use ckks_he::random_distributions::generate_random_vector;
use ckks_he::keys::client_key::{calculate_error, to_plaintext};

let (mut client_key, server_key) = generate_pair_keys_toy();
let plaintext_dimension = 1 << (client_key.dimension_exponent() -1);
let plaintext_bound = 5.0;
let real_plaintext = generate_random_vector(
        plaintext_dimension,
        -plaintext_bound,
        plaintext_bound,
);
let plaintext = to_plaintext(&real_plaintext);
let ciphertext = client_key.encrypt(&plaintext, plaintext_bound).unwrap();
let decrypted = client_key.decrypt(&ciphertext);
let error = calculate_error(&plaintext, &decrypted);
assert!(error < client_key.rescaled_error(&ciphertext), "Error large than expected");

pub fn rescaled_error(&self, ct: &Ciphertext<T>) -> f64

Returns an upper bound of the error expected after decryption.

pub fn rescaled_upperbound_message(&self, ct: &Ciphertext<T>) -> f64

Returns an upper bound of the message expected after decryption.

pub fn dimension_exponent(&self) -> u32

Returns h such that 2^h is the dimension of the cyclotomic ring of messages.

pub fn plaintext_dimension(&self) -> usize

Returns the dimension of plaintexts space.